+++ sdk/recoveryservices/azure-resourcemanager-recoveryservices/src/main/java/com/azure/resourcemanager/recoveryservices/implementation/RecoveryServicesManagementClientImpl.java
@@ -15,6 +15,7 @@ import com.azure.core.management.exception.ManagementException;
 import com.azure.core.management.polling.PollResult;
 import com.azure.core.management.polling.PollerFactory;
 import com.azure.core.util.Context;
+import com.azure.core.util.CoreUtils;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.polling.AsyncPollResponse;
 import com.azure.core.util.polling.LongRunningOperationStatus;
@@ -37,15 +38,12 @@ import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
-import java.util.Map;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 /** Initializes a new instance of the RecoveryServicesManagementClientImpl type. */
 @ServiceClient(builder = RecoveryServicesManagementClientBuilder.class)
 public final class RecoveryServicesManagementClientImpl implements RecoveryServicesManagementClient {
-    private final ClientLogger logger = new ClientLogger(RecoveryServicesManagementClientImpl.class);
-
     /** The subscription Id. */
     private final String subscriptionId;
 
@@ -276,10 +274,7 @@ public final class RecoveryServicesManagementClientImpl implements RecoveryServi
      * @return the merged context.
      */
     public Context mergeContext(Context context) {
-        for (Map.Entry<Object, Object> entry : this.getContext().getValues().entrySet()) {
-            context = context.addData(entry.getKey(), entry.getValue());
-        }
-        return context;
+        return CoreUtils.mergeContexts(this.getContext(), context);
     }
 
     /**
@@ -342,8 +337,8 @@ public final class RecoveryServicesManagementClientImpl implements RecoveryServi
                         if (managementError.getCode() == null || managementError.getMessage() == null) {
                             managementError = null;
                         }
-                    } catch (IOException ioe) {
-                        logger.logThrowableAsWarning(ioe);
+                    } catch (IOException | RuntimeException ioe) {
+                        LOGGER.logThrowableAsWarning(ioe);
                     }
                 }
             } else {
@@ -371,7 +366,7 @@ public final class RecoveryServicesManagementClientImpl implements RecoveryServi
             super(null);
             this.statusCode = statusCode;
             this.httpHeaders = httpHeaders;
-            this.responseBody = responseBody.getBytes(StandardCharsets.UTF_8);
+            this.responseBody = responseBody == null ? null : responseBody.getBytes(StandardCharsets.UTF_8);
         }
 
         public int getStatusCode() {
@@ -402,4 +397,6 @@ public final class RecoveryServicesManagementClientImpl implements RecoveryServi
             return Mono.just(new String(responseBody, charset));
         }
     }
+
+    private static final ClientLogger LOGGER = new ClientLogger(RecoveryServicesManagementClientImpl.class);
 }