+++ sdk/recoveryservices/azure-resourcemanager-recoveryservices/CHANGELOG.md
@@ -1,7 +1,23 @@
 # Release History
 
-## 1.0.0-beta.2 (Unreleased)
+## 1.0.0-beta.2 (2022-07-19)
 
+- Azure Resource Manager RecoveryServices client library for Java. This package contains Microsoft Azure SDK for RecoveryServices Management SDK. Recovery Services Client. Package tag package-2016-06. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
+
+### Features Added
+
+#### `RecoveryServicesManager$Configurable` was modified
+
+* `withRetryOptions(com.azure.core.http.policy.RetryOptions)` was added
+* `withScope(java.lang.String)` was added
+
+#### `RecoveryServicesManager` was modified
+
+* `authenticate(com.azure.core.http.HttpPipeline,com.azure.core.management.profile.AzureProfile)` was added
+
+#### `models.Vault` was modified
+
+* `resourceGroupName()` was added
 
 ## 1.0.0-beta.1 (2020-12-18)